Fix likelihoodField-bug index calculation and data types

+++ ike_localization/include/ike_localization/mcl/likelihoodField.hpp
@@ -31,7 +31,7 @@ public:
   double resolution_;         // 受け取ったマップの解像度
   double origin_x_;           // 受け取ったマップの原点x
   double origin_y_;           // 受け取ったマップの原点y
-  std::vector<int8_t> data_;  // 受け取ったマップの各画素の情報
+  std::vector<double> data_;  // 受け取ったマップの各画素の情報
 };
 }  // namespace mcl
